refactor: simplify node search internals for reviewability
- Keep only NODE_SEARCH_MIN_QUERY_LENGTH in @n8n/api-types (the one
  constant the frontend shares); result/candidate caps move into the
  service that uses them
- Drop unused sharing option fields from getManyWithSharingSubquery
- Drop unreachable try/catch around JSON.stringify of DB-deserialised
  parameters, inline the single-use scopes constant
- Extract one shared corpus seeder for the three perf files
- Top-level createOwner import in repository test
