+++ packages/@n8n/api-types/src/dto/index.ts
@@ -142,10 +142,6 @@ export {
 export {
 	SearchWorkflowNodesQueryDto,
 	NODE_SEARCH_MIN_QUERY_LENGTH,
-	NODE_SEARCH_MAX_QUERY_LENGTH,
-	NODE_SEARCH_MAX_RESULTS,
-	NODE_SEARCH_PER_WORKFLOW_CAP,
-	NODE_SEARCH_MAX_CANDIDATE_WORKFLOWS,
 	type NodeSearchHit,
 	type SearchWorkflowNodesResponse,
 } from './workflows/search-workflow-nodes.dto';
