+++ packages/cli/test/performance/node-search-postgres-plan.perf.ts
@@ -15,15 +15,12 @@
  */
 import { testDb } from '@n8n/backend-test-utils';
 import { GlobalConfig } from '@n8n/config';
-import { ProjectRepository, SharedWorkflowRepository, WorkflowRepository } from '@n8n/db';
 import { Container } from '@n8n/di';
 import { DataSource } from '@n8n/typeorm';
-import { v4 as uuid } from 'uuid';
 
-import { createOwner } from '../integration/shared/db/users';
+import { seedCorpus } from './shared';
 
 const CORPUS = 20_000;
-const LOREM = 'onboarding flow reads from the CRM and forwards it to billing. '.repeat(6);
 
 const isPostgres = () => Container.get(GlobalConfig).database.type === 'postgresdb';
 
@@ -36,53 +33,7 @@ describe('postgres planner hint', () => {
 	beforeAll(async () => {
 		if (!isPostgres()) return;
 		await testDb.init();
-		await testDb.truncate([
-			'SharedWorkflow',
-			'ProjectRelation',
-			'WorkflowEntity',
-			'Project',
-			'User',
-		]);
-		const owner = await createOwner();
-		const project = await Container.get(ProjectRepository).getPersonalProjectForUserOrFail(
-			owner.id,
-		);
-		const repo = Container.get(WorkflowRepository);
-		const shared = Container.get(SharedWorkflowRepository);
-
-		for (let start = 0; start < CORPUS; start += 500) {
-			const size = Math.min(500, CORPUS - start);
-			const rows = Array.from({ length: size }, (_, k) =>
-				repo.create({
-					id: `pgp-${(start + k).toString().padStart(7, '0')}`,
-					name: `WF ${start + k}`,
-					active: false,
-					isArchived: false,
-					nodes: [
-						{
-							id: uuid(),
-							name: `Step 3 of flow ${start + k}`,
-							type: 'n8n-nodes-base.set',
-							typeVersion: 1,
-							position: [0, 0] as [number, number],
-							parameters: { body: LOREM },
-						},
-					],
-					connections: {},
-					nodeGroups: [],
-					versionId: uuid(),
-					settings: {},
-				}),
-			);
-			await repo.insert(rows);
-			await shared.insert(
-				rows.map((w) => ({
-					workflowId: w.id,
-					projectId: project.id,
-					role: 'workflow:owner' as const,
-				})),
-			);
-		}
+		await seedCorpus(CORPUS);
 
 		ds = Container.get(DataSource);
 		// Match a real instance, where autovacuum has long since analysed.
